ENH: Reorganizing classification after UNC.
At the UNC visit we looked closer at how to classify the tissue
parts. In particular, we looked at a multi-pass approach for
classifying first the tubules, and then revisiting them one by
one.

The modifications in this patch are a first step towards creating
a multi-pass classification approach. We want to be able to first
run a classification for wall (ephitelium) and the inside of the
tubule.